refactor: вынести функции из компонента App
