+++ src/components/Register/Register.js
@@ -89,10 +89,10 @@ const Register = (props) => {
         ? ''
         : Object.entries(errors)
           .map(([errKey, errValue]) => (
-            <ErrorMessage key={errKey} text={errValue} />
+            <ErrorMessage key={errKey} text={errValue} isErrorVisible={true} />
           ))}
 
-      <ErrorMessage text={props.errorText} />
+      <ErrorMessage text={props.errorText} isErrorVisible={props.isErrorVisible} />
 
       <input
         type="submit"
