+++ src/utils/MainApi.js
@@ -103,8 +103,8 @@ class Api {
 }
 
 const api = new Api({
-  // baseUrl: '//api.movies-explorer.sun.nomoredomains.monster',
-  baseUrl: 'http://localhost:3000',
+  baseUrl: '//api.movies-explorer.sun.nomoredomains.monster',
+  // baseUrl: 'http://localhost:3000',
 });
 
 export default api;
